Fix problems with log refnum -1 in /set log_rewrite.

+++ include/log.h
@@ -12,9 +12,9 @@
 	FILE	*do_log 	(int, const char *, FILE **);
 	void	logger 		(const void *);
 	void	set_log_file 	(const void *);
-	void	add_to_log 	(FILE *, unsigned, const unsigned char *, int, const char *);
+	void	add_to_log 	(FILE *, long, const unsigned char *, int, const char *);
 	BUILT_IN_COMMAND(logcmd);
-	void	add_to_logs	(int, int, const char *, int, const char *);
+	void	add_to_logs	(long, int, const char *, int, const char *);
 	char *	logctl		(char *);
 
 #endif /* _LOG_H_ */
